move store page to reserved path
